+++ packages/nx-firebase/src/generators/function/lib/add-function-config.ts
