+++ packages/nx-firebase/src/generators/function/lib/index.ts
@@ -1,3 +1,3 @@
-export * from './add-function'
+export * from './add-function-config'
 export * from './create-files'
 export * from './update-project'
